refactor: refactor config package and made a bit cleaner
