+++ src/main/java/com/quiz_app/service/auth/AuthenticationService.java
@@ -1,6 +1,6 @@
 package com.quiz_app.service.auth;
 
-import com.quiz_app.config.JwtService;
+import com.quiz_app.config.jwt.JwtService;
 import com.quiz_app.controller.authcontroller.*;
 import com.quiz_app.entity.jwttoken.Token;
 import com.quiz_app.entity.jwttoken.TokenType;
